requirements.txt
Django
psycopg2
ipython
django-extensions
python-dotenv
djangorestframework
drf-nested-routers
